requirements.txt
# Core
streamlit==1.32.0
numpy==1.24.3
pillow==10.2.0
pandas==2.0.3

# Audio Analysis
gTTS==2.5.0
SpeechRecognition==3.10.1
pydub==0.25.1

# Image Analysis - OpenCV
opencv-python-headless==4.9.0.80
opencv-contrib-python-headless==4.9.0.80

# Image Analysis - AI/ML
deepface==0.0.87
rembg==2.0.55

# Deep Learning Backends
onnxruntime==1.17.0
tensorflow==2.15.0
tf-keras==2.15.0
protobuf==3.20.3

# Text Analysis
textblob==0.18.0
nltk==3.8.1
